Move back the Json digitization stuff to the Json module
