+++ Examples/Algorithms/Digitization/CMakeLists.txt
@@ -4,8 +4,7 @@ add_library(
   src/MeasurementCreation.cpp
   src/PlanarSteppingAlgorithm.cpp
   src/SmearingAlgorithm.cpp
-  src/DigitizationOptions.cpp
-  src/JsonDigitizationConfig.cpp)
+  src/DigitizationOptions.cpp)
 target_include_directories(
   ActsExamplesDigitization
   PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>)
@@ -13,7 +12,7 @@ target_link_libraries(
   ActsExamplesDigitization
   PRIVATE
     ActsCore ActsPluginDigitization ActsPluginIdentification
-    ActsExamplesFramework ActsExamplesIoJson
+    ActsExamplesFramework
     Boost::program_options)
 
 install(
