+++ Examples/Io/Json/CMakeLists.txt
@@ -2,13 +2,14 @@ add_library(
   ActsExamplesIoJson SHARED
   src/JsonMaterialWriter.cpp
   src/JsonSurfacesWriter.cpp
-  src/JsonOptionsWriter.cpp)
+  src/JsonOptionsWriter.cpp
+  src/JsonDigitizationConfig.cpp)
 target_include_directories(
   ActsExamplesIoJson
   PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>)
 target_link_libraries(
   ActsExamplesIoJson
-  PUBLIC ActsCore ActsPluginJson ActsExamplesFramework Boost::program_options)
+  PUBLIC ActsCore ActsPluginJson ActsExamplesDigitization ActsExamplesFramework Boost::program_options)
 
 install(
   TARGETS ActsExamplesIoJson
